requirements.txt
absl-py==1.0.0
aioredis==1.3.1
asgiref==3.2.10
astor==0.8.0
astroid==2.3.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==22.2.0
autobahn==22.12.1
Automat==0.8.0
beautifulsoup4==4.11.1
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.1
channels==3.0.1
channels-redis==3.2.0
chardet==3.0.4
ckiptagger==0.0.20
click==8.1.3
colorama==0.4.1
constantly==15.1.0
cryptography==38.0.4
daphne==3.0.2
defusedxml==0.6.0
diff-match-patch==20181111
dill==0.3.1.1
Django==3.0.7
django-cors-headers==3.1.1
django-import-export==2.0.2
django-templated-mail==1.1.1
djangorestframework==3.10.3
djoser==2.0.3
et-xmlfile==1.0.1
filelock==3.12.2
flatbuffers==1.12
fsspec==2023.1.0
future==0.18.1
fuzzywuzzy==0.17.0
gast==0.2.2
gensim==4.2.0
gevent==1.4.0
google-auth==1.6.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
googleapis-common-protos==1.6.0
greenlet==0.4.15
grpcio==1.34.0
grpcio-tools==1.34.0
gunicorn==20.1.0
h11==0.14.0
h5py==2.10.0
hiredis==1.0.0
huggingface-hub==0.16.4
hyperlink==21.0.0
idna==2.8
importlib-metadata==5.2.0
incremental==22.10.0
isort==4.3.21
jdcal==1.4.1
jieba==0.39
joblib==1.1.0
keras==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
libclang==14.0.1
Markdown==3.1.1
MarkupPy==1.14
mccabe==0.6.1
msgpack==1.0.4
numpy==1.21.6
oauthlib==3.1.0
odfpy==1.4.1
opencc-python-reimplemented==0.1.6
openpyxl==3.0.3
opt-einsum==3.1.0
packaging==21.3
pandas==1.3.5
promise==2.2.1
protobuf==3.10.0
psycopg2-binary==2.8.4
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==1.9.0
pylint==2.4.3
pyOpenSSL==22.1.0
pyparsing==3.0.9
pypinyin==0.36.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2019.3
PyYAML==5.3
redis==3.5.3
regex==2023.6.3
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
safetensors==0.3.1
scikit-learn==1.0.2
scipy==1.7.3
service-identity==21.1.0
six==1.12.0
sklearn==0.0
SkPy==0.10.4
smart-open==6.0.0
soupsieve==2.3.2.post1
sqlparse==0.3.0
tablib==1.1.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-datasets==1.3.0
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==0.15.1
tensorflow-text==2.9.0
termcolor==1.1.0
threadpoolctl==3.1.0
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.30.2
Twisted==22.10.0
txaio==22.2.1
typed-ast==1.4.0
typing_extensions==4.2.0
urllib3==1.25.6
uvicorn==0.20.0
websocket==0.2.1
websocket-client==0.56.0
websockets==10.4
Werkzeug==2.1.2
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
zhconv==1.4.0
zipp==3.11.0
zope.interface==4.6.0
